Refactor by creating a module to list item
