+++ src/book_operations.rb
@@ -1,9 +1,12 @@
 require './src/book'
+require './src/modules/list'
 
 class BookOerations
     def initialize(book)
         @books = book
     end
+
+    include List
 
     def create_book
         print 'Title: '
@@ -16,13 +19,5 @@ class BookOerations
     puts 'Book created successfully'  
     end
 
-    def list_all_books
-        puts 'There are no books yet. Kindly Create books' if @books.empty?
-        puts ' ====================================================================================='
-        p
-        @books.each do |book|
-          puts "Title: #{book.title}, Author: #{book.author}"
-        end
-        puts ' ====================================================================================='
-      end
+    
 end
